+++ Source/Engine/Graphics/OpenGL/RenderContext.h
@@ -15,7 +15,7 @@ public:
 
 	void Initialize();
 	void CreateBuffers(std::vector<Model>& aModels);
-	void Render(const std::vector<Model>& aModels, const TextureLibrary& aTextureLibrary, ShaderLibrary& aShaderLibrary, int aWidth, int aHeight, double aDeltaTime);
+	void Render(const std::vector<Model>& aModels, const TextureLibrary& aTextureLibrary, ShaderLibrary& aShaderLibrary, int aWidth, int aHeight, float aDeltaTime);
 	void Destroy(const std::vector<Model>& aModels);
 
 private:
